Added bone animation
Only proof of concept, real implementation will be better

+++ lib/src/data/animation_bone_frame.dart
@@ -11,7 +11,7 @@ class AnimationBoneFrame {
   // (1, 2]  :ease in out
 
   final int duration;
-  final int tweenEasing;
+  final double tweenEasing;
   final Transform transform;
 
   AnimationBoneFrame(
