+++ lib/src/data/animation_slot_frame.dart
@@ -11,7 +11,7 @@ class AnimationSlotFrame {
   // (1, 2]  :ease in out
 
   final int duration;
-  final int tweenEasing;
+  final double tweenEasing;
   final int displayIndex;
   final int zOrder;
   final ColorTransform colorTransform;
